refactor UserInput to Argument
merge conflict
