+++ src/Web/Hyperbole.hs
@@ -181,7 +181,8 @@ import Network.Wai.Handler.Warp as Warp (run)
 import Web.Atomic.CSS ()
 import Web.Atomic.Types ()
 import Web.Hyperbole.Application
-import Web.Hyperbole.Data.Encoded (FromEncoded, ToEncoded, UserInput (..))
+import Web.Hyperbole.Data.Argument (UserInput (..))
+import Web.Hyperbole.Data.Encoded (FromEncoded, ToEncoded)
 import Web.Hyperbole.Data.QueryData
 import Web.Hyperbole.Document
 import Web.Hyperbole.Effect.Client
