+++ src/Web/Hyperbole/HyperView/Input.hs
@@ -7,8 +7,7 @@ import Data.String.Conversions (cs)
 import Data.Text (Text)
 import GHC.Generics (Generic)
 import Web.Atomic.Types
-import Web.Hyperbole.Data.Argument (encodeArgument)
-import Web.Hyperbole.Data.Encoded (UserInput)
+import Web.Hyperbole.Data.Argument (UserInput, encodeArgument)
 import Web.Hyperbole.HyperView.Event (DelayMs, onChange, onClick, onInput)
 import Web.Hyperbole.HyperView.Types (HyperView (..))
 import Web.Hyperbole.Route (Route (..), routeUri)
